7/20 Features page
Added a new route to get the name for the chair page(tested and working), merged, settings and features page to one Features page.

Edited the features page. and created the ajax running at an interval, using the arduino url to retreive data.(needs to change abit regarding the ports that the sensor is using and also the if else statement in these ajax statement)